food_delivery/requirements.txt
# 项目说明、安装与运行步骤见同目录下的 README.md
Django>=5.0,<6
djangorestframework>=3.15,<4
